+++ CurrencySpender/Helpers/PlayerHelper.cs
@@ -1,8 +1,10 @@
 using CurrencySpender.Data;
 using FFXIVClientStructs.FFXIV.Client.Game.UI;
+using FFXIVClientStructs.FFXIV.Client.System.Framework;
 using FFXIVClientStructs.FFXIV.Client.UI;
 using FFXIVClientStructs.FFXIV.Client.UI.Agent;
 using FFXIVClientStructs.FFXIV.Component.GUI;
+using System.Text;
 
 namespace CurrencySpender.Helpers
 {